Passwords Change, Profile Pic, Account Activities
- Added Links to SignUp and SignIn to link to each other
- Added Lost Password link on SignIn => ForgottenPassword => send email => Auth Handler => confirm password change
- Added password change to profile
- Migrate Profile Picture Change to a FormInputFile
- Migrate LoginAttempt to AccountActivity
- For this, added Signout, Signup, Password Change and Password Lost to account activities
- Added an enum for AccountActivity types
- Added the declined status for roles : user that requested the declined role can't request it, but managers can still offer them a previsouly declined role
- Removed the Chat system (component, classes, service and dashboard tile) that will be migrated into real-time
- Optimized the brands fetching that was not used in some views
- Corrected a bug in RoleAdd and RoleOffer due to the change of behaviour on the Choose component
- Migrate the getGETparam to the Utils Service (previsouly used only for Tabs)
- Corrected a bug due to a fetch of the id of employee, not stored anymore in employee (fetched from the current user)
- Removed the stat called in functions as well as on the dashboard
- Added the denied roles list to the Roles component
- Created the Auth directory for component to separate auth-related forms from other forms
- Corrected a behaviour on ExTable that would have been prevented to load items if items are null after loading have been put to false
- Corrected a behaviour on Choose that would have been prevented to load items if selection is null (and not empty array or empty string)

+++ src/components/App/Roles/Roles.js
@@ -36,17 +36,20 @@ const Roles = () => {
   const [requestedRoles, setRequestedRoles] = useState({});
   const [requestedRolesEmployees, setRequestedRolesEmployees] = useState({});
 
+  const [userDeniedRoles, setUserDeniedRoles] = useState({});
+  const [userDeniedRolesCompanies, setUserDeniedRolesCompanies] = useState({});
+
   const observerKey = uuid();
   
   const [computed, setComputed] = useState(DataService.computed.getDefaultComputedValues());
   
   const computeRoles = () => {
-    if(!computed.user) return;
+    if(!computed.user) { return; }
     // USER ROLES
-    RoleService.getRolesForEmployeeId(computed.user.uid, [ERoleStatus.CONFIRMED, ERoleStatus.DRAFT, ERoleStatus.REVOKED])
+    RoleService.getRolesForEmployeeId(computed.user.uid, [ERoleStatus.CONFIRMED, ERoleStatus.DRAFT, ERoleStatus.REVOKED, ERoleStatus.DENIED])
       .then(roles => {
-        let userRoles = {}, userRevokedRoles = {}, userDraftRoles = {}, 
-          userRolesCompanyKeys = [], userRevokedRolesCompanyKeys = [], userDraftRolesCompanyKeys = [];
+        let userRoles = {}, userRevokedRoles = {}, userDraftRoles = {}, userDeniedRoles = {},
+          userRolesCompanyKeys = [], userRevokedRolesCompanyKeys = [], userDraftRolesCompanyKeys = [], userDeniedRolesCompanyKeys = [];
 
         Object.keys(roles).forEach(roleKey => {
           if(roles[roleKey].status === ERoleStatus.CONFIRMED) {
@@ -61,30 +64,37 @@ const Roles = () => {
             userRevokedRoles[roleKey] = roles[roleKey];
             userRevokedRolesCompanyKeys.push(roles[roleKey].companyId);
           }
+          if(roles[roleKey].status === ERoleStatus.DENIED) {
+            userDeniedRoles[roleKey] = roles[roleKey];
+            userDeniedRolesCompanyKeys.push(roles[roleKey].companyId);
+          }
         });
 
         CompanyService.getAllForIdList(UtilsService.removeDuplicateFromArray(Object.keys(roles).map(roleKey => roles[roleKey].companyId)))
           .then(allCompanies => {
-            let userRolesCompanies = {}, userDraftRolesCompanies = {}, userRevokedRolesCompanies = {};
+            let userRolesCompanies = {}, userDraftRolesCompanies = {}, userRevokedRolesCompanies = {}, userDeniedRolesCompanies = {};
 
             userRolesCompanyKeys.forEach(companyKey => userRolesCompanies[companyKey] = allCompanies[companyKey]);
             userRevokedRolesCompanyKeys.forEach(companyKey => userRevokedRolesCompanies[companyKey] = allCompanies[companyKey]);
             userDraftRolesCompanyKeys.forEach(companyKey => userDraftRolesCompanies[companyKey] = allCompanies[companyKey]);
+            userDeniedRolesCompanyKeys.forEach(companyKey => userDeniedRolesCompanies[companyKey] = allCompanies[companyKey]);
 
             setUserDraftRolesCompanies(userDraftRolesCompanies);
             setUserRevokedRolesCompanies(userRevokedRolesCompanies);
             setUserRolesCompanies(userRolesCompanies);
+            setUserDeniedRolesCompanies(userDeniedRolesCompanies);
 
             setUserDraftRoles(userDraftRoles);
             setUserRevokedRoles(userRevokedRoles);
             setUserRoles(userRoles);
+            setUserDeniedRoles(userDeniedRoles);
           })
           .catch(ErrorService.manageError);
       })
       .catch(ErrorService.manageError);
 
     // REQUESTED ROLES
-    if(computed.activeRole) {
+    if(computed.activeRole && computed.activeRole.role === ERole.MANAGER) {
       RoleService.getRolesForCompanyId(computed.activeRole.companyId, [ERoleStatus.DRAFT])
         .then(requestedRoles => {
           let employeesIds = UtilsService.removeDuplicateFromArray(Object.keys(requestedRoles).map(roleKey => requestedRoles[roleKey].employeeId));
@@ -116,11 +126,15 @@ const Roles = () => {
     return defaultActions;
   };
 
-  useEffect(() => computeRoles(), [computed]);
+  useEffect(() => {
+    if(computed.initialized) {
+      computeRoles();
+    }
+  }, [computed]);
 
   useEffect(() => {
     DataService.computed.observeComputedValues(setComputed, observerKey);
-    return () => DataService.computed.unobserveComputedValues(observerKey)
+    return () => DataService.computed.unobserveComputedValues(observerKey);
   }, []);
   
   if(!computed.initialized) { return null; }
@@ -156,6 +170,13 @@ const Roles = () => {
       options={ {showDraft: true, showActions: true} } />
   );
 
+  const renderDeniedUserRole = (itemKey, itemData) => (
+    <RoleCompany key={itemKey} 
+      company={ {[itemKey]: itemData} } 
+      roles={UtilsService.filterKeyValueOnPropertyValue(userDeniedRoles, predicate => predicate.companyId === itemKey)}
+      options={ {showActions: true, showDates: true} } />
+  );
+
   return (
     <div className="Roles">
       {computed.activeRole && 
@@ -185,6 +206,9 @@ const Roles = () => {
 
       <h1>Revoked Roles</h1>
       <ExTable items={userRevokedRolesCompanies} renderItem={renderRevokedUserRole} isNoFrame />
+
+      <h1>Denied Roles</h1>
+      <ExTable items={userDeniedRolesCompanies} renderItem={renderDeniedUserRole} isNoFrame />
 
       <ActionButton icon={<Icon source="fa" icon={faPlus} />} actions={computeActions()} />
     </div>
